requirements.txt
# Core ML
torch>=2.2.0
torchvision

# Flash Attention 2 (install separately with: pip install flash-attn --no-build-isolation)
# flash-attn>=2.5.0

# HuggingFace ecosystem
transformers>=4.40.0
datasets>=2.19.0
huggingface_hub>=0.23.0
safetensors>=0.4.0

# Tokenizer
tiktoken>=0.7.0

# Training utilities
wandb>=0.17.0

# Data processing
numpy>=1.26.0
tqdm>=4.66.0

# Config
pyyaml>=6.0.1
